Implement "Delete Empty Cards" action in Deck Picker Compose UI
- Updated `MoreOptionsMenuActions` and `DeckPickerActions` to remove the unused `Long` parameter from the `onDeleteEmptyCards` callback.
- Connected the "Delete Empty Cards" menu item in `DeckPickerScreen` to the navigation actions.
- Implemented the `onDeleteEmptyCards` action in `DeckPickerNavHost` to display the `EmptyCardsDialogFragment`.

+++ AnkiDroid/src/main/java/com/ichi2/anki/deckpicker/compose/DeckPickerActions.kt
@@ -42,7 +42,7 @@ data class MoreOptionsMenuActions(
     val onCheckDatabase: () -> Unit,
     val onExport: () -> Unit,
     val onStartStudy: () -> Unit,
-    val onDeleteEmptyCards: (Long) -> Unit,
+    val onDeleteEmptyCards: () -> Unit,
 )
 
 /**
